Revert "Code cleaning"
This reverts commit 6e1036224aca7c2c5efc5b46c8c3f67e509f5929.

+++ include/FileCore.h
@@ -27,7 +27,7 @@ namespace FileCore {
 
 bool loadConfig();
 bool saveConfig();
-void getFSInfo(int8_t);
+void getFSInfo(int);
 } // namespace FileCore
 } // namespace ESParaSite
